requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
eventlet>=0.33.0
python-chess>=1.999
gunicorn>=20.0.4